Make business types, investment stages and types read-only in admin
Also fix CH company field order in admin.
